Add files via upload
This is a shopping app in IT project.
The shopping application project verifies the feasibility of basic web technologies to build practical shopping applications, constructs a simple and efficient online shopping platform, focuses on user experience and core functions, and makes up for the shortcomings of existing solutions. Despite the limited functionality, it achieves the core objectives and lays the foundation for subsequent optimization. In the future, we can integrate back-end storage data, user personalization and other advanced functions to make the application more complete. The team members also improved their front-end development and problem solving skills.